chore: revamp chess example
