+++ chess/src/stackr/machine.ts
