+++ chess/src/stackr/mru.ts
@@ -1,8 +1,8 @@
 import { MicroRollup } from "@stackr/sdk";
 import { stackrConfig } from "../../stackr.config.ts";
 
-import { moveSchema } from "./actions.ts";
-import { chessStateMachine } from "./chess.machine.ts";
+import { moveSchema } from "./schemas.ts";
+import { chessStateMachine } from "./machine.ts";
 
 const mru = await MicroRollup({
   config: stackrConfig,
