+++ chess/src/stackr/schemas.ts
