+++ chess/tests/chess.test.ts
@@ -7,8 +7,8 @@ import { StateMachine } from "@stackr/sdk/machine";
 import { expect } from "chai";
 import { Chess } from "chess.js";
 import genesisState from "../genesis-state.json";
-import { moveSchema } from "../src/stackr/actions.ts";
-import { chessStateMachine } from "../src/stackr/chess.machine.ts";
+import { chessStateMachine } from "../src/stackr/machine.ts";
+import { moveSchema } from "../src/stackr/schemas.ts";
 import { ChessState } from "../src/stackr/state.ts";
 import { transitions } from "../src/stackr/transitions.ts";
 import { signByOperator, sleep } from "../src/utils.ts";
